[REF] l10n_be_vat_reports: remove partner.vat
