+++ l10n_be_vat_reports/tests/test_vat_reports.py
@@ -3,9 +3,6 @@
 
 from odoo import fields
 
-from lxml.etree import XML
-import base64
-
 from .common import TestVatReportsCommon
 
 
@@ -16,30 +13,16 @@ class TestVatListing(TestVatReportsCommon):
         tax = self.env.ref("l10n_be.%s_attn_VAT-OUT-21-S" % company.id)
         self._create_test_data(tax)
 
-    def _prepare_listing(self):
+    def test_partner_vat_listing(self):
         year = fields.Date.today().year
-        wizard = self.env["partner.vat"].create(
-            {"year": str(year), "limit_amount": 1}
+        partner_vat_list = self.env["partner.vat.list"].create(
+            {"year": year, "limit_amount": 1}
         )
-        action = wizard.get_partners()
-        vat_list_id = action["res_id"]
-        vat_list = self.env["partner.vat.list"].browse(vat_list_id)
+        partner_vat_list.get_partners()
 
-        self.assertEqual(94.5, vat_list.total_vat)
-        return vat_list
+        self.assertEquals(partner_vat_list.total_turnover, 450.)
+        self.assertEqual(partner_vat_list.total_vat, 94.5)
 
-    def test_xml_list(self):
-        full_list = self._prepare_listing()
-        full_list.create_xml()
-        ns = {"ns2": "http://www.minfin.fgov.be/ClientListingConsignment"}
-        xml = XML(base64.b64decode(full_list.file_save))
-        xml_vat_amount = xml.xpath(
-            '//ns2:Client[ns2:CompanyVATNumber[text() = "0477472701"]]'
-            "/ns2:VATAmount",
-            namespaces=ns,
-        )[0].text
-        self.assertEqual("94.50", xml_vat_amount)
-
-    def test_pdf_list(self):
-        full_list = self._prepare_listing()
-        full_list.print_vatlist()
+        # coverage
+        partner_vat_list.create_xml()
+        partner_vat_list.print_vatlist()
