+++ l10n_be_vat_reports/wizard/l10n_be_partner_vat_listing.xml
@@ -10,9 +10,9 @@
 
         <record id="view_partner_vat_listing" model="ir.ui.view">
             <field name="name">Partner VAT Listing</field>
-            <field name="model">partner.vat</field>
+            <field name="model">partner.vat.list</field>
             <field name="arch" type="xml">
-                <form string="Partner VAT Listing">
+                <form>
                     <p class="oe_grey">
                         This wizard will create an XML file for VAT details and total invoiced amounts per partner.
                     </p>
@@ -29,18 +29,20 @@
             </field>
         </record>
 
-        <record id="action_partner_vat_listing" model="ir.actions.act_window">
+        <record id="partner_vat_list_action_parameters" model="ir.actions.act_window">
             <field name="name">Partner VAT Listing</field>
             <field name="type">ir.actions.act_window</field>
-            <field name="res_model">partner.vat</field>
+            <field name="res_model">partner.vat.list</field>
             <field name="view_type">form</field>
             <field name="view_mode">form</field>
             <field name="view_id" ref="view_partner_vat_listing"/>
             <field name="target">new</field>
         </record>
 
-        <menuitem name="Annual Listing Of VAT-Subjected Customers" parent="l10n_be.account_reports_be_statements_menu"
-                  action="action_partner_vat_listing" id="partner_vat_listing"/>
+        <menuitem name="Annual Listing Of VAT-Subjected Customers"
+                  parent="l10n_be.account_reports_be_statements_menu"
+                  action="partner_vat_list_action_parameters"
+                  id="partner_vat_list_menu"/>
 
         <record id="view_vat_listing" model="ir.ui.view">
             <field name="name">Vat Listing</field>
@@ -65,19 +67,6 @@
                             <button colspan="2" name="print_vatlist" string="Print" type="object" icon="fa-print"/>
                         </group>
                     </group>
-                </form>
-            </field>
-        </record>
-
-        <record id="view_vat_listing_result" model="ir.ui.view">
-            <field name="name">VAT List - XML File Created</field>
-            <field name="model">partner.vat.list</field>
-            <field name="priority" eval="1"/>
-            <field name="arch" type="xml">
-                <form string="Save XML">
-                    <separator string="Note: "/>
-                    <label string="Save the File :" for="file_save"/>
-                    <field name="file_save" filename="name"/>
                 </form>
             </field>
         </record>
